requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.1
django-rest-framework==0.1.0
djangorestframework==3.13.0
djangorestframework-jwt==1.11.0
psycopg2==2.9.2
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
